# docs/api/api-spec.yaml
openapi: 3.0.3
info:
  title: Dcard Lite
  version: 1.0.0
  description: Based on rp-001-register-login
servers:
  - url: http://localhost:8080
    description: Local server

tags:
  - name: Auth
    description: Register & Login
  - name: Boards
    description: List boards

paths:
  /users/register:
    post:
      tags: [Auth]
      summary: User register
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            examples:
              example:
                value:
                  name: "Leo"
                  email: "leo@example.com"
                  password: "abc12345"
                  confirmPassword: "abc12345"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
              examples:
                success:
                  value:
                    userId: 123
                    displayName: "Leo"
                    email: "leo@example.com"
                    role: "USER"
                    createdAt: "2025-12-25T10:00:00Z"
        "400":
          description: Bad Request (validation failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                nameInvalid:
                  value:
                    status: 400
                    error: "Bad Request"
                    message: "VALIDATION_FAILED"
                    code: "NAME_INVALID"
                    path: "/users/register"
                    timestamp: "2025-12-25T10:00:00Z"
                emailInvalid:
                  value:
                    status: 400
                    error: "Bad Request"
                    message: "VALIDATION_FAILED"
                    code: "EMAIL_INVALID"
                    path: "/users/register"
                    timestamp: "2025-12-25T10:00:00Z"
                passwordInvalid:
                  value:
                    status: 400
                    error: "Bad Request"
                    message: "VALIDATION_FAILED"
                    code: "PASSWORD_INVALID"
                    path: "/users/register"
                    timestamp: "2025-12-25T10:00:00Z"
                confirmPasswordInvalid:
                  value:
                    status: 400
                    error: "Bad Request"
                    message: "VALIDATION_FAILED"
                    code: "CONFIRM_PASSWORD_INVALID"
                    path: "/users/register"
                    timestamp: "2025-12-25T10:00:00Z"
        "409":
          description: Conflict (email already exists)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                emailAlreadyExists:
                  value:
                    status: 409
                    error: "Conflict"
                    message: "CONFLICT"
                    code: "EMAIL_ALREADY_EXISTS"
                    path: "/users/register"
                    timestamp: "2025-12-25T10:00:00Z"

  /users/login:
    post:
      tags: [Auth]
      summary: User login
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              example:
                value:
                  email: "leo@example.com"
                  password: "abc12345"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
              examples:
                success:
                  value:
                    userId: 123
                    displayName: "Leo"
                    role: "USER"
                    accessToken: "..."
        "400":
          description: Bad Request (validation failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                emailInvalid:
                  value:
                    status: 400
                    error: "Bad Request"
                    message: "VALIDATION_FAILED"
                    code: "EMAIL_INVALID"
                    path: "/users/login"
                    timestamp: "2025-12-25T10:00:00Z"
                passwordInvalid:
                  value:
                    status: 400
                    error: "Bad Request"
                    message: "VALIDATION_FAILED"
                    code: "PASSWORD_INVALID"
                    path: "/users/login"
                    timestamp: "2025-12-25T10:00:00Z"
        "401":
          description: Unauthorized (authentication failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                authenticationFailed:
                  value:
                    status: 401
                    error: "Unauthorized"
                    message: "UNAUTHORIZED"
                    code: "AUTHENTICATION_FAILED"
                    path: "/users/login"
                    timestamp: "2025-12-25T10:00:00Z"

  /boards:
    get:
      tags: [Boards]
      summary: List boards
      operationId: listBoards
      parameters:
        - in: query
          name: page
          required: false
          description: Page number (default=1, must be >=1)
          schema:
            type: integer
            format: int32
            default: 1
            minimum: 1
          example: 1
        - in: query
          name: pageSize
          required: false
          description: Page size (default=20, 1..100)
          schema:
            type: integer
            format: int32
            default: 20
            minimum: 1
            maximum: 100
          example: 20
        - in: query
          name: keyword
          required: false
          description: Optional keyword for board name fuzzy search (blank treated as not provided; validated after trim 1..50)
          schema:
            type: string
            maxLength: 50
          example: "八卦"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListBoardsResponse"
              examples:
                success:
                  value:
                    page: 1
                    pageSize: 20
                    total: 2
                    items:
                      - boardId: 1
                        name: "八卦版"
                        description: "想聊什麼就聊什麼"
                      - boardId: 2
                        name: "軟體版"
                        description: "聊軟體相關的知識"
                empty:
                  value:
                    page: 1
                    pageSize: 20
                    total: 0
                    items: []
        "400":
          description: Bad Request (validation failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                pageInvalid:
                  value:
                    status: 400
                    error: "Bad Request"
                    message: "VALIDATION_FAILED"
                    code: "PAGE_INVALID"
                    path: "/boards"
                    timestamp: "2025-12-25T10:00:00Z"
                pageSizeInvalid:
                  value:
                    status: 400
                    error: "Bad Request"
                    message: "VALIDATION_FAILED"
                    code: "PAGE_SIZE_INVALID"
                    path: "/boards"
                    timestamp: "2025-12-25T10:00:00Z"
                keywordInvalid:
                  value:
                    status: 400
                    error: "Bad Request"
                    message: "VALIDATION_FAILED"
                    code: "KEYWORD_INVALID"
                    path: "/boards"
                    timestamp: "2025-12-25T10:00:00Z"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                internalError:
                  value:
                    status: 500
                    error: "Internal Server Error"
                    message: "INTERNAL_ERROR"
                    code: "INTERNAL_ERROR"
                    path: "/boards"
                    timestamp: "2025-12-25T10:00:00Z"

  /boards/{boardId}/posts:
    post:
      tags: [Boards]
      summary: Create a post in a specific board
      security:
        - bearerAuth: []
      operationId: createPost
      parameters:
        - in: path
          name: boardId
          required: true
          description: The ID of the board where the post will be created
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
            examples:
              example:
                value:
                  title: "關於SpringBoot的問題"
                  body: "請問怎麼建立一個SpringBoot Maven專案"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePostResponse"
              examples:
                success:
                  value:
                    postId: 1
        "400":
          description: Bad Request (validation failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                titleInvalid:
                  value:
                    status: 400
                    error: "Bad Request"
                    message: "VALIDATION_FAILED"
                    code: "TITLE_INVALID"
                    path: "/boards/2/posts"
                    timestamp: "2025-12-25T10:00:00Z"
                bodyInvalid:
                  value:
                    status: 400
                    error: "Bad Request"
                    message: "VALIDATION_FAILED"
                    code: "BODY_INVALID"
                    path: "/boards/2/posts"
                    timestamp: "2025-12-25T10:00:00Z"
        "401":
          description: Unauthorized (authentication failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  value:
                    status: 401
                    error: "Unauthorized"
                    message: "UNAUTHORIZED"
                    code: "SECURITY_UNAUTHORIZED"
                    path: "/boards/2/posts"
                    timestamp: "2025-12-25T10:00:00Z"
        "404":
          description: Not Found (board does not exist)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                boardNotFound:
                  value:
                    status: 404
                    error: "Not Found"
                    message: "NOT_FOUND"
                    code: "BOARD_NOT_FOUND"
                    path: "/boards/3/posts"
                    timestamp: "2025-12-25T10:00:00Z"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                internalError:
                  value:
                    status: 500
                    error: "Internal Server Error"
                    message: "INTERNAL_ERROR"
                    code: "UNEXPECTED_ERROR"
                    path: "/boards/2/posts"
                    timestamp: "2025-12-25T10:00:00Z"

components:
  securitySchemes:
      bearerAuth:            # 這個名稱可以自訂，稍後套用時會用到
        type: http
        scheme: bearer
        bearerFormat: JWT    # 提示使用者這是 JWT 格式
  schemas:
    RegisterRequest:
      type: object
      required: [name, email, password, confirmPassword]
      properties:
        name:
          type: string
          description: User display name (validated after trim)
          maxLength: 20
          example: "Leo"
        email:
          type: string
          description: Email (validated after trim, stored lowercase)
          format: email
          maxLength: 100
          example: "leo@example.com"
        password:
          type: string
          description: Password (8-12 chars, at least one letter and one digit; validated after trim)
          minLength: 8
          maxLength: 12
          example: "abc12345"
        confirmPassword:
          type: string
          description: Must match password (validated after trim)
          minLength: 8
          maxLength: 12
          example: "abc12345"

    RegisterResponse:
      type: object
      required: [userId, displayName, email, role, createdAt]
      properties:
        userId:
          type: integer
          format: int64
          example: 123
        displayName:
          type: string
          maxLength: 20
          example: "Leo"
        email:
          type: string
          format: email
          maxLength: 100
          example: "leo@example.com"
        role:
          type: string
          example: "USER"
        createdAt:
          type: string
          format: date-time
          description: ISO-8601 UTC
          example: "2025-12-25T10:00:00Z"

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 100
          description: Email (validated after trim)
          example: "leo@example.com"
        password:
          type: string
          minLength: 8
          maxLength: 12
          description: Password (validated after trim)
          example: "abc12345"

    LoginResponse:
      type: object
      required: [userId, displayName, role, accessToken]
      properties:
        userId:
          type: integer
          format: int64
          example: 123
        displayName:
          type: string
          maxLength: 20
          example: "Leo"
        role:
          type: string
          example: "USER"
        accessToken:
          type: string
          description: JWT access token
          example: "..."

    BoardItem:
      type: object
      required: [boardId, name, description]
      properties:
        boardId:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          maxLength: 20
          example: "八卦版"
        description:
          type: string
          maxLength: 100
          example: "想聊什麼就聊什麼"

    ListBoardsResponse:
      type: object
      required: [page, pageSize, total, items]
      properties:
        page:
          type: integer
          format: int32
          example: 1
        pageSize:
          type: integer
          format: int32
          example: 20
        total:
          type: integer
          format: int64
          example: 2
        items:
          type: array
          items:
            $ref: "#/components/schemas/BoardItem"

    CreatePostRequest:
      type: object
      required: [title, body]
      properties:
        title:
          type: string
          description: Post title (validated after trim, no `<` or `>`)
          maxLength: 50
          example: "關於SpringBoot的問題"
        body:
          type: string
          description: Post body (validated after trim, allows `\n`, no `<` or `>`)
          maxLength: 300
          example: "請問怎麼建立一個SpringBoot Maven專案"

    CreatePostResponse:
      type: object
      required: [postId]
      properties:
        postId:
          type: integer
          format: int64
          example: 1

    ErrorResponse:
      type: object
      required: [status, error, message, code, path, timestamp]
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          description: HTTP reason phrase (framework-generated)
          example: "Bad Request"
        message:
          type: string
          description: Error category
          enum:
            - VALIDATION_FAILED
            - UNAUTHORIZED
            - CONFLICT
          example: "VALIDATION_FAILED"
        code:
          type: string
          description: Error code detail
          example: "EMAIL_INVALID"
        path:
          type: string
          example: "/users/register"
        timestamp:
          type: string
          format: date-time
          example: "2025-12-25T10:00:00Z"